+++ src/modules/PreviewObjectColors.cpp
@@ -1,7 +1,10 @@
 #include "PreviewObjectColors.hpp"
 #include "ScrollableObjects.hpp"
 #include <alphalaneous.level-storage-api/include/LevelStorageAPI.hpp>
+#include <alphalaneous.editortab_api/include/EditorTabAPI.hpp>
+#include "modules/UIScaling.hpp"
 #include "utils/Constants.hpp"
+#include "utils/Utils.hpp"
 
 void PreviewObjectColors::onEditor() {
     auto savedObj = alpha::level_storage::getSavedValue<std::string>(m_editorLayer, "color-object");
@@ -25,30 +28,63 @@ void PreviewObjectColors::onEditor() {
 
     m_editorUI->schedule(schedule_selector(POCEditorUI::updateObjectColors));
 
-    onUpdateButtons();
-
-    addEventListener(UpdateButtonsEvent(), [this] {
-        onUpdateButtons();
-    });
-
-    m_editorUI->runAction(CallFuncExt::create([this, fields] {
-        if (ScrollableObjects::isEnabled()) {
+    if (ScrollableObjects::isEnabled()) {
+        m_editorUI->runAction(CallFuncExt::create([this] {
             for (auto child : m_editorUI->getChildrenExt()) {
                 if (auto bar = typeinfo_cast<EditButtonBar*>(child)) {
                     if (!bar->m_hasCreateItems) continue;
                     auto soBar = static_cast<SOEditButtonBar*>(bar);
-                    auto btn = CCMenuItemExt::createSpriteExtraWithFrameName("GJ_editHSVBtn2_001.png", 1, [this, fields] (CCMenuItemSpriteExtra* sender) {
-                        auto customizeObjectLayer = CustomizeObjectLayer::create(fields->m_defaultObject, nullptr);
-                        customizeObjectLayer->show();
-                        customizeObjectLayer->updateColorSprite();
-                        customizeObjectLayer->updateChannelLabel(fields->m_defaultObject->m_baseColor->m_colorID);
+                    auto btn = CCMenuItemExt::createSpriteExtraWithFrameName("GJ_editHSVBtn2_001.png", 1.f, [this] (CCMenuItemSpriteExtra* sender) {
+                        editColor();
                     });
                     btn->setZOrder(-1);
+                    btn->setID("color-preview"_spr);
                     soBar->addToExtrasMenu(btn);
                 }
             }
-        }
-    }));
+        }));
+    }
+    else {
+        auto btnContainer = CCNode::create();
+        btnContainer->setID("color-preview-container"_spr);
+        btnContainer->setZOrder(100);
+        btnContainer->setAnchorPoint({1.f, 1.f});
+        m_editorUI->addChild(btnContainer);
+        m_editorUI->m_uiItems->addObject(btnContainer);
+
+        auto btn = geode::Button::createWithSpriteFrameName("GJ_editHSVBtn2_001.png", [this] (auto sender) {
+            editColor();
+            float scale = 1.f;
+            if (UIScaling::isEnabled()) {
+                scale = UIScaling::get()->m_scale;
+            }
+        });
+        btn->setID("color-preview"_spr);
+        btnContainer->setContentSize(btn->getScaledContentSize());
+        btn->setPosition(btnContainer->getContentSize() / 2.f);
+
+        btnContainer->addChild(btn);
+
+        addEventListener(UIScaleUpdated(), [this, btn, btnContainer] (float scale, bool scaleToolbars, bool fullReload) {
+            auto winSize = CCDirector::get()->getWinSize();
+            float x = winSize.width - (95.f * scale);
+            btnContainer->setScale(0.65f * scale);
+
+            auto spacer = m_editorUI->getChildByID("spacer-line-right");
+            if (spacer) {
+                x = spacer->boundingBox().getMinX();
+            }
+            
+            btnContainer->setPosition({x - 3.f * scale, tinker::utils::getToolbarHeight() - 4.f * scale});
+        });
+
+        alpha::editor_tabs::addTabSwitchCallback([this, btn] (auto id) {
+            setButtonVisible(btn);
+        });
+        alpha::editor_tabs::addModeSwitchCallback([this, btn] (auto id) {
+            setButtonVisible(btn);
+        });
+    }
 
     addEventListener(LevelSavedEvent(), [this] {
         auto editorUI = static_cast<POCEditorUI*>(m_editorUI);
@@ -57,25 +93,25 @@ void PreviewObjectColors::onEditor() {
     });
 }
 
-void PreviewObjectColors::onUpdateButtons() {
-    if (!ScrollableObjects::isEnabled()) {
-        m_editorUI->m_editObjectBtn->setOpacity(255);
-        m_editorUI->m_editObjectBtn->setColor({255, 255, 255});
-        m_editorUI->m_editObjectBtn->m_animationEnabled = true;
-    }
+void PreviewObjectColors::setButtonScale(geode::Button* button, float scale) {
 }
 
-void POCEditorUI::editObject(cocos2d::CCObject* sender) {
-    if (ScrollableObjects::isEnabled()) return EditorUI::editObject(sender);
-    auto fields = m_fields.self();
+void PreviewObjectColors::setButtonVisible(geode::Button* button) {
+    auto tabRes = alpha::editor_tabs::nodeForTab(alpha::editor_tabs::getCurrentTab().unwrapOrDefault());
+    if (!tabRes) return;
+    auto tab = typeinfo_cast<EditButtonBar*>(tabRes.unwrap().data());
 
-    if (!m_selectedObject && (!m_selectedObjects || m_selectedObjects->count() == 0)) {
-        auto customizeObjectLayer = CustomizeObjectLayer::create(fields->m_defaultObject, nullptr);
-        customizeObjectLayer->show();
-    }
-    else {
-        EditorUI::editObject(sender);
-    }
+    button->setVisible(tab && tab->m_tabIndex != 13 && tab->m_hasCreateItems && alpha::editor_tabs::getCurrentMode().unwrapOrDefault() == alpha::editor_tabs::BUILD);
+}
+
+void PreviewObjectColors::editColor() {
+    auto editorUI = static_cast<POCEditorUI*>(m_editorUI);
+    auto fields = editorUI->m_fields.self();
+
+    auto customizeObjectLayer = CustomizeObjectLayer::create(fields->m_defaultObject, nullptr);
+    customizeObjectLayer->show();
+    customizeObjectLayer->updateColorSprite();
+    customizeObjectLayer->updateChannelLabel(fields->m_defaultObject->m_baseColor->m_colorID);
 }
 
 GameObject* POCEditorUI::createObject(int objectID, cocos2d::CCPoint position) {