+++ src/settings/BoolSettingNode.cpp
@@ -1,0 +1,15 @@
+#include <Geode/Geode.hpp>
+#include "SettingNode.hpp"
+
+using namespace geode::prelude;
+
+class BoolSettingNode : public tinker::settings::SettingNode<BoolSettingNode, "boolA"> {
+public:
+
+    bool init(float width, std::shared_ptr<SettingV3> setting) {
+        if (!SettingNodeBase::init(width, setting)) return false;
+
+
+        return true;
+    }
+};