+++ src/settings/SettingsData.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+
+
+class SettingsData {
+
+    geode::SettingV3* m_setting;
+
+
+};